requirements.txt
dotenv
streamlit
google-generativeai